# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce Analytics API
  description: API for accessing ecommerce analytics data including sales, orders, products, and transactions
  version: 1.0.0
  contact:
    name: API Support
    email: support@ecommerce-analytics.com

servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.ecommerce-analytics.com/v1
    description: Production server

tags:
  - name: Sales
    description: Sales and revenue data
  - name: Orders
    description: Order information and statistics
  - name: Products
    description: Product performance data
  - name: Transactions
    description: Transaction history

paths:
  /sales/monthly:
    get:
      tags:
        - Sales
      summary: Get monthly sales data
      description: Returns revenue and order count for each month
      operationId: getMonthlySales
      parameters:
        - name: year
          in: query
          description: Year to filter sales data
          required: false
          schema:
            type: integer
            example: 2024
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  months:
                    type: array
                    items:
                      type: string
                    example: ["January", "February", "March"]
                  revenue:
                    type: array
                    items:
                      type: number
                    example: [45000, 52000, 48000]
                  orders:
                    type: array
                    items:
                      type: integer
                    example: [450, 520, 480]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /sales/statistics:
    get:
      tags:
        - Sales
      summary: Get sales statistics
      description: Returns aggregated sales statistics including total revenue, orders, and averages
      operationId: getSalesStatistics
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalRevenue:
                    type: number
                    example: 812000
                  totalOrders:
                    type: integer
                    example: 8120
                  totalProducts:
                    type: integer
                    example: 4120
                  avgOrderValue:
                    type: number
                    example: 100.00
                  growthRate:
                    type: number
                    example: 12.5

  /sales/categories:
    get:
      tags:
        - Sales
      summary: Get sales by category
      description: Returns sales breakdown by product category
      operationId: getSalesByCategory
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: object
                    additionalProperties:
                      type: number
                    example:
                      Electronics: 285000
                      Clothing: 198000
                      "Home & Garden": 156000

  /products/top:
    get:
      tags:
        - Products
      summary: Get top selling products
      description: Returns list of top selling products by revenue
      operationId: getTopProducts
      parameters:
        - name: limit
          in: query
          description: Number of products to return
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Successful response

  /products/{productId}/price:
    put:
      tags:
        - Products
      summary: Update product price
      description: Updates the price for a specific product. Can be used to increase or decrease prices.
      operationId: updateProductPrice
      parameters:
        - name: productId
          in: path
          description: Product ID
          required: true
          schema:
            type: string
            example: "prod-001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - price
              properties:
                price:
                  type: number
                  description: New price for the product
                  example: 79.99
                  minimum: 0
      responses:
        '200':
          description: Price updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Product price updated successfully"
                  product:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "prod-001"
                      name:
                        type: string
                        example: "Wireless Headphones"
                      category:
                        type: string
                        example: "Electronics"
                      oldPrice:
                        type: number
                        example: 99.99
                      newPrice:
                        type: number
                        example: 79.99
                      stock:
                        type: integer
                        example: 150
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product details
      description: Returns detailed information about a specific product
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: Product ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders/recent:
    get:
      tags:
        - Orders
      summary: Get recent orders
      description: Returns list of recent orders
      operationId: getRecentOrders
      parameters:
        - name: limit
          in: query
          description: Number of orders to return
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: Filter by order status
          required: false
          schema:
            type: string
            enum: [completed, pending, processing, cancelled]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: Returns detailed information about a specific order
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /transactions:
    get:
      tags:
        - Transactions
      summary: Get transaction history
      description: Returns paginated list of transactions
      operationId: getTransactions
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of transactions per page
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: startDate
          in: query
          description: Filter transactions from this date
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter transactions until this date
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: "prod-001"
        name:
          type: string
          example: "Wireless Headphones"
        sales:
          type: number
          example: 45000
        units:
          type: integer
          example: 450
        category:
          type: string
          example: "Electronics"

    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            description:
              type: string
              example: "High-quality wireless headphones with noise cancellation"
            price:
              type: number
              example: 99.99
            stock:
              type: integer
              example: 150
            rating:
              type: number
              example: 4.5

    Order:
      type: object
      properties:
        id:
          type: string
          example: "ORD-1234"
        date:
          type: string
          format: date
          example: "2024-12-15"
        customer:
          type: string
          example: "John Smith"
        product:
          type: string
          example: "Wireless Headphones"
        amount:
          type: number
          example: 99.99
        status:
          type: string
          enum: [completed, pending, processing, cancelled]
          example: "completed"

    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            customerId:
              type: string
              example: "cust-5678"
            productId:
              type: string
              example: "prod-001"
            quantity:
              type: integer
              example: 1
            shippingAddress:
              type: string
              example: "123 Main St, City, State 12345"
            paymentMethod:
              type: string
              example: "Credit Card"

    Transaction:
      type: object
      properties:
        id:
          type: string
          example: "txn-9876"
        orderId:
          type: string
          example: "ORD-1234"
        date:
          type: string
          format: date-time
          example: "2024-12-15T10:30:00Z"
        amount:
          type: number
          example: 99.99
        type:
          type: string
          enum: [sale, refund, adjustment]
          example: "sale"
        status:
          type: string
          enum: [completed, pending, failed]
          example: "completed"

    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8

    Error:
      type: object
      properties:
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Invalid parameter value"
        details:
          type: string
          example: "Year must be a valid 4-digit number"

# Made with Bob